refactor: remove dead code & split modules units
