+++ valuator/runtime.py
@@ -4,14 +4,13 @@ import json
 from datetime import datetime
 from typing import Any
 
-from .session.store import markdown_document, render_report_markdown
+from .session.markdown import render_final_markdown
 from .utils.time_utils import utc_isoformat
 
 
 def create_tool_registry(model: str, usage_writer: Any | None = None):
     from .tools.base import ToolRegistry
     from .tools.code_execute_tool import ExecuteCodeTool
-    # from .tools.domain_tool import DomainTool
     from .tools.sec_tool import SECTool
     from .tools.web_search_tool import PerplexitySearchTool
     from .tools.yfinance_tool import YFinanceBalanceSheetTool
@@ -24,7 +23,6 @@ def create_tool_registry(model: str, usage_writer: Any | None = None):
         code_tool,
         YFinanceBalanceSheetTool(),
         SECTool(model=model),
-        # DomainTool(model=model),
     ):
         registry.register(tool)
     registry.bind_usage_writer(usage_writer)
@@ -32,16 +30,9 @@ def create_tool_registry(model: str, usage_writer: Any | None = None):
 
 
 def final_output_text(output: Any) -> str:
-    if isinstance(output, str):
-        return output.strip()
-    if isinstance(output, dict):
-        for key in ("markdown", "report", "content"):
-            value = output.get(key)
-            if isinstance(value, str) and value.strip():
-                return value.strip()
-    rendered = render_report_markdown(output)
+    rendered = render_final_markdown(output)
     if rendered:
-        return markdown_document("Final", rendered)
+        return rendered
     return json.dumps(output, ensure_ascii=False, indent=2, default=str).strip()
