+++ valuator/session/store.py
@@ -16,19 +16,15 @@ from valuator.utils.config import ROOT_DIR
 from valuator.utils.time_utils import utc_isoformat
 
 from .browse_tree import build_browse_tree
+from .markdown import (
+    render_final_markdown,
+    strip_markdown_title,
+)
+from .report_artifacts import artifact_text, load_report_source, render_aggregation_report
+from .task_tree import build_task_tree_snapshot, render_tree_markdown
 from .trace import SessionTraceWriter, task_rel_path
 
 CURRENT_ROUND = 1
-PRIMARY_MARKDOWN_KEYS = ("markdown", "report", "content")
-REPORT_MARKDOWN_KEYS = (
-    "markdown",
-    "report",
-    "content",
-    "findings",
-    "result",
-    "summary",
-    "domain_summary",
-)
 
 
 def valuator_sessions_root() -> Path:
@@ -37,91 +33,6 @@ def valuator_sessions_root() -> Path:
 
 def round_dir_name(round_number: int) -> str:
     return f"round-{round_number:02d}"
-
-
-def extract_primary_markdown_text(value: Any) -> str:
-    if isinstance(value, str):
-        return value.strip()
-    if isinstance(value, Mapping):
-        for key in PRIMARY_MARKDOWN_KEYS:
-            candidate = value.get(key)
-            if isinstance(candidate, str) and candidate.strip():
-                return candidate.strip()
-    return ""
-
-
-def strip_markdown_title(markdown: str) -> str:
-    text = markdown.strip()
-    if not text.startswith("#"):
-        return text
-    lines = text.splitlines()
-    if not lines or not lines[0].startswith("#"):
-        return text
-    index = 1
-    while index < len(lines) and not lines[index].strip():
-        index += 1
-    return "\n".join(lines[index:]).strip()
-
-
-def markdown_document(title: str, body: str) -> str:
-    content = strip_markdown_title(body).strip()
-    if not content:
-        return ""
-    return f"# {title}\n\n{content}"
-
-
-def render_report_markdown(value: Any) -> str:
-    if isinstance(value, Mapping) and value.get("status") == "facts_only":
-        facts = value.get("facts")
-        value = facts if isinstance(facts, Mapping) else facts
-
-    if isinstance(value, Mapping):
-        for key in REPORT_MARKDOWN_KEYS:
-            candidate = value.get(key)
-            if isinstance(candidate, str) and candidate.strip():
-                return strip_markdown_title(candidate)
-
-    text = extract_primary_markdown_text(value)
-    if text:
-        return strip_markdown_title(text)
-
-    def render(current: Any, heading_level: int = 2) -> str:
-        if current is None:
-            return ""
-        if isinstance(current, str):
-            return current.strip()
-        if isinstance(current, Mapping):
-            lines: list[str] = []
-            for key, item in current.items():
-                if key in {"status", "source_task_id", "query", "sources"}:
-                    continue
-                rendered = render(item, min(heading_level + 1, 6)).strip()
-                if not rendered:
-                    continue
-                label = str(key).replace("_", " ").strip()
-                if "\n" not in rendered and not rendered.startswith("- "):
-                    lines.append(f"- **{label}**: {rendered}")
-                else:
-                    lines.extend([f"{'#' * heading_level} {label}", "", rendered, ""])
-            return "\n".join(lines).strip()
-        if isinstance(current, list):
-            if all(not isinstance(item, (Mapping, list)) for item in current):
-                return "\n".join(
-                    f"- {text}"
-                    for text in (str(item).strip() for item in current)
-                    if text
-                ).strip()
-            lines = []
-            for index, item in enumerate(current, start=1):
-                rendered = render(item, min(heading_level + 1, 6)).strip()
-                if rendered:
-                    lines.extend(
-                        [f"{'#' * heading_level} Item {index}", "", rendered, ""]
-                    )
-            return "\n".join(lines).strip()
-        return str(current).strip()
-
-    return render(value).strip()
 
 
 class ValuatorSessionStore:
@@ -237,7 +148,6 @@ class ValuatorSessionStore:
         with self._lock:
             self._root_task_id = root_task.id
             self._session_payload["root_task_id"] = root_task.id
-            self._plan_tasks.clear()
 
             unit_count = 0
             default_domain_id = ""
@@ -247,15 +157,23 @@ class ValuatorSessionStore:
                 if domain_ids:
                     default_domain_id = domain_ids[0]
 
-            tree = self._write_task_tree(
+            snapshot = build_task_tree_snapshot(
                 root_task,
+                root_task_id=root_task.id,
                 unit_count=unit_count,
                 default_domain_id=default_domain_id,
+                session_dir=self.session_dir,
+                tasks_dir=self.tasks_dir,
+                task_created_at=self._task_created_at,
+                task_artifacts=self._task_artifacts,
             )
-            self._write_json(self.plan_active_dir / "task_tree.json", tree)
+            self._plan_tasks = snapshot.plan_tasks
+            for task_id, payload in snapshot.task_payloads.items():
+                self._write_json(self._task_dir(task_id) / "task.json", payload)
+            self._write_json(self.plan_active_dir / "task_tree.json", snapshot.tree)
             self._write_text(
                 self.plan_active_dir / "task_tree.md",
-                f"# Task Tree\n\n{self._render_tree_md(root_task)}\n",
+                f"# Task Tree\n\n{snapshot.tree_markdown}\n",
             )
 
             # The server seeds the root task before query analysis exists.
@@ -320,7 +238,7 @@ class ValuatorSessionStore:
                 domain_summary = result.result.strip()
 
             self._write_text(
-                result_path, self._artifact_text(result.result, result.error)
+                result_path, artifact_text(result.result, result.error)
             )
             self._write_json(
                 result_json_path,
@@ -367,55 +285,32 @@ class ValuatorSessionStore:
                 list(plan_task["deps"]) if plan_task["task_type"] == "merge" else []
             )
             child_sources = [
-                self._load_report_source(child_task_id)
+                load_report_source(
+                    task_id=child_task_id,
+                    task_dir=self._task_dir(child_task_id),
+                    task_name=str(
+                        self._plan_tasks.get(child_task_id, {}).get("task_name")
+                        or child_task_id
+                    ).replace("_", " "),
+                    rel=self._rel,
+                )
                 for child_task_id in source_reports
             ]
-            current_source = self._load_report_source(task_id)
-
-            title = str(plan_task.get("task_name") or task_id).replace("_", " ")
-            lines = [f"# {title}", ""]
-            summary = render_report_markdown(output)
-            if not summary:
-                summary = current_source["markdown"] or render_report_markdown(
-                    current_source["report_input"]
-                )
-            if summary:
-                lines.extend([summary, ""])
-
-            evidence_started = False
-            for child in child_sources:
-                body = child["markdown"] or render_report_markdown(
-                    child["report_input"]
-                )
-                if not body:
-                    continue
-                if summary and not evidence_started:
-                    lines.extend(["## Supporting Evidence", ""])
-                    evidence_started = True
-                lines.extend([f"### {child['title']}", "", body, ""])
-
-            if not summary and not evidence_started:
-                lines = [f"# {title}", "", "(no report content)"]
-            self._write_text(
-                report_path,
-                "\n".join(lines).strip() + "\n",
+            current_source = load_report_source(
+                task_id=task_id,
+                task_dir=self._task_dir(task_id),
+                task_name=str(plan_task.get("task_name") or task_id).replace("_", " "),
+                rel=self._rel,
             )
 
-            if isinstance(output, Mapping):
-                facts = (
-                    dict(output["facts"])
-                    if output.get("status") == "facts_only"
-                    and isinstance(output.get("facts"), Mapping)
-                    else dict(output)
-                )
-            elif isinstance(output, str) and output.strip():
-                facts = {"summary": output.strip()}
-            else:
-                facts = (
-                    current_source["raw_result"]
-                    if current_source["raw_result"] is not None
-                    else {}
-                )
+            title = str(plan_task.get("task_name") or task_id).replace("_", " ")
+            report_markdown, facts = render_aggregation_report(
+                title=title,
+                output=output,
+                child_sources=child_sources,
+                current_source=current_source,
+            )
+            self._write_text(report_path, report_markdown)
 
             domain_ids: list[str] = []
             if self._analysis_payload is not None:
@@ -431,12 +326,12 @@ class ValuatorSessionStore:
                     "source_task_ids": source_reports,
                     "child_results": [
                         {
-                            "task_id": child["task_id"],
-                            "title": child["title"],
-                            "source_type": child["source_type"],
-                            "report_path": child["report_path"],
-                            "raw_results_path": child["raw_path"],
-                            "raw_result": child["raw_result"],
+                            "task_id": child.task_id,
+                            "title": child.title,
+                            "source_type": child.source_type,
+                            "report_path": child.report_path,
+                            "raw_results_path": child.raw_path,
+                            "raw_result": child.raw_result,
                         }
                         for child in child_sources
                     ],
@@ -456,10 +351,6 @@ class ValuatorSessionStore:
             self._write_session()
 
     def final_output_markdown(self, output: Any) -> str:
-        primary = extract_primary_markdown_text(output)
-        if primary:
-            return primary
-
         with self._lock:
             report_body = ""
             if self._root_task_id is not None:
@@ -470,14 +361,7 @@ class ValuatorSessionStore:
                     report_body = strip_markdown_title(
                         report_path.read_text(encoding="utf-8")
                     )
-
-        if report_body:
-            return markdown_document("Final", report_body)
-
-        rendered = render_report_markdown(output)
-        if rendered:
-            return markdown_document("Final", rendered)
-        return ""
+        return render_final_markdown(output, report_body=report_body)
 
     def write_final_output(self, content: Any, root_task: Task | None = None) -> None:
         with self._lock:
@@ -533,7 +417,7 @@ class ValuatorSessionStore:
             if root_task is not None:
                 self._write_text(
                     self.output_dir / "final.trace.md",
-                    f"# Final Trace\n\n{self._render_tree_md(root_task)}\n",
+                    f"# Final Trace\n\n{render_tree_markdown(root_task)}\n",
                 )
             self._session_payload["updated_at"] = utc_isoformat()
             self._write_session()
@@ -572,85 +456,6 @@ class ValuatorSessionStore:
             else:
                 self._session_payload.pop("error", None)
             self._write_session()
-
-    def _write_task_tree(
-        self,
-        task: Task,
-        *,
-        unit_count: int,
-        default_domain_id: str,
-    ) -> dict[str, Any]:
-        created_at = self._task_created_at.setdefault(task.id, utc_isoformat())
-        task_dir = self._task_dir(task.id)
-        children = task.children()
-        task_type = "merge" if task.id == self._root_task_id or children else "leaf"
-        query_unit_ids = list(task.query_unit_ids)
-        if not query_unit_ids:
-            query_unit_ids = (
-                list(range(unit_count))
-                if task.id == self._root_task_id
-                else [0] if task_type == "leaf" and unit_count == 1 else []
-            )
-        tool = None
-        if task_type == "leaf":
-            if task.last_tool_request is not None:
-                tool = {
-                    "name": task.last_tool_request.tool_name,
-                    "args": dict(task.last_tool_request.args),
-                }
-            elif task.tool_hint.strip():
-                tool = {"name": task.tool_hint.strip(), "args": {}}
-
-        plan_task = {
-            "id": task.id,
-            "task_type": task_type,
-            "query_unit_ids": query_unit_ids,
-            "deps": [child.id for child in children] if task_type == "merge" else [],
-            "tool": tool,
-            "domain_id": default_domain_id,
-            "output": "execution/result.md" if task_type == "leaf" else "",
-            "description": task.description,
-            "task_name": task.task_name,
-            "merge_instruction": "",
-        }
-        self._plan_tasks[task.id] = plan_task
-
-        payload = {
-            "task_id": task.id,
-            "parent_id": task.parent_id,
-            "description": task.description,
-            "task_name": task.task_name,
-            "state": task.state.value,
-            "error": task.error,
-            "created_at": created_at,
-            "updated_at": utc_isoformat(),
-            "task_type": task_type,
-            "query_unit_ids": query_unit_ids,
-            "deps": list(plan_task["deps"]),
-            "tool": tool,
-            "domain_id": default_domain_id,
-            "output": plan_task["output"],
-            "merge_instruction": "",
-            "children": [child.id for child in children],
-            "artifacts": self._artifacts_for(task.id),
-        }
-        self._write_json(task_dir / "task.json", payload)
-
-        children_payload = [
-            self._write_task_tree(
-                child,
-                unit_count=unit_count,
-                default_domain_id=default_domain_id,
-            )
-            for child in children
-        ]
-        return {
-            "task_id": task.id,
-            "description": task.description,
-            "state": task.state.value,
-            "path": self._rel(task_dir),
-            "children": children_payload,
-        }
 
     def build_browse_tree(self) -> None:
         with self._lock:
@@ -706,106 +511,6 @@ class ValuatorSessionStore:
         return hashlib.sha256(text.encode("utf-8")).hexdigest()
 
     @staticmethod
-    def _artifact_text(value: Any, error: str | None) -> str:
-        if isinstance(value, str) and value.strip():
-            return value.strip() + "\n"
-        if isinstance(value, dict):
-            for key in ("markdown", "report", "content", "findings", "result"):
-                candidate = value.get(key)
-                if isinstance(candidate, str) and candidate.strip():
-                    return candidate.strip() + "\n"
-        if error:
-            return error.strip() + "\n"
-        return (
-            json.dumps(value, ensure_ascii=False, indent=2, default=str).strip() + "\n"
-        )
-
-    def _load_report_source(self, task_id: str) -> dict[str, Any]:
-        task_dir = self._task_dir(task_id)
-        task_name = str(
-            self._plan_tasks.get(task_id, {}).get("task_name") or task_id
-        ).replace("_", " ")
-        candidates = [
-            (
-                "aggregation",
-                task_dir / "aggregation" / "report.md",
-                task_dir / "aggregation" / "raw_results.json",
-            ),
-            (
-                "execution",
-                task_dir / "execution" / "result.md",
-                task_dir / "execution" / "result.json",
-            ),
-        ]
-
-        for source_type, markdown_path, raw_path in candidates:
-            raw_payload = self._read_json_if_exists(raw_path)
-            report_input = None
-            raw_result = None
-            if raw_payload is not None:
-                if source_type == "aggregation":
-                    report_input = raw_payload.get("output")
-                    if report_input is None:
-                        report_input = raw_payload.get("facts")
-                    raw_result = raw_payload.get("facts")
-                    if raw_result is None:
-                        raw_result = raw_payload.get("output")
-                else:
-                    report_input = raw_payload.get("raw_result")
-                    raw_result = raw_payload.get("raw_result")
-            if markdown_path.exists():
-                return {
-                    "task_id": task_id,
-                    "title": task_name,
-                    "source_type": source_type,
-                    "markdown": strip_markdown_title(
-                        markdown_path.read_text(encoding="utf-8")
-                    ),
-                    "report_input": report_input,
-                    "raw_result": raw_result,
-                    "report_path": self._rel(markdown_path),
-                    "raw_path": (
-                        self._rel(raw_path) if raw_payload is not None else None
-                    ),
-                }
-            if raw_payload is not None:
-                return {
-                    "task_id": task_id,
-                    "title": task_name,
-                    "source_type": source_type,
-                    "markdown": "",
-                    "report_input": report_input,
-                    "raw_result": raw_result,
-                    "report_path": None,
-                    "raw_path": self._rel(raw_path),
-                }
-
-        return {
-            "task_id": task_id,
-            "title": task_name,
-            "source_type": "",
-            "markdown": "",
-            "report_input": None,
-            "raw_result": None,
-            "report_path": None,
-            "raw_path": None,
-        }
-
-    @staticmethod
-    def _render_report_markdown(value: Any) -> str:
-        return render_report_markdown(value)
-
-    @staticmethod
-    def _strip_markdown_title(markdown: str) -> str:
-        return strip_markdown_title(markdown)
-
-    @staticmethod
-    def _read_json_if_exists(path: Path) -> dict[str, Any] | None:
-        if not path.exists():
-            return None
-        return json.loads(path.read_text(encoding="utf-8"))
-
-    @staticmethod
     def _write_text(path: Path, content: str) -> None:
         path.parent.mkdir(parents=True, exist_ok=True)
         path.write_text(content, encoding="utf-8")
@@ -817,20 +522,3 @@ class ValuatorSessionStore:
             json.dumps(payload, ensure_ascii=False, indent=2, default=str) + "\n",
             encoding="utf-8",
         )
-
-    @staticmethod
-    def _render_tree_md(task: Task, depth: int = 0) -> str:
-        indent = "  " * depth
-        children = task.children()
-        kind = (
-            f"leaf: {task.tool_hint}" if not children and task.tool_hint.strip() else ""
-        )
-        suffix = f" ({kind})" if kind else ""
-        name_snippet = (
-            f" `{task.task_name.strip()}`" if str(task.task_name or "").strip() else ""
-        )
-        line = f"{indent}- **{task.id}**{name_snippet} [{task.state.value}] {task.description}{suffix}"
-        lines = [line]
-        for child in children:
-            lines.append(ValuatorSessionStore._render_tree_md(child, depth + 1))
-        return "\n".join(lines)
