+++ valuator/tools/specs.py
@@ -179,55 +179,6 @@ TOOL_SPECS: dict[str, ToolSpec] = {
             "timeout": {"type": "integer"},
         },
     ),
-    # "domain_tool": ToolSpec(
-    #     name="domain_tool",
-    #     optional=(
-    #         "corp",
-    #         "company_name",
-    #         "ticker",
-    #         "query",
-    #         "context",
-    #         "grounding_mode",
-    #         "as_of_utc",
-    #         "time_scope",
-    #         "target_start",
-    #         "target_end",
-    #         "domain_guide",
-    #         "domain_persona",
-    #         "domain_rubric",
-    #         "domain_format",
-    #         "domain_id",
-    #     ),
-    #     capability="aspect-guided domain analysis via persona/rubric/format",
-    #     llm_required=(),
-    #     param_descriptions={
-    #         "corp": "Company identifier",
-    #         "company_name": "Company legal name",
-    #         "ticker": "Listing ticker",
-    #         "query": "Task-specific query",
-    #         "context": "Contextual background",
-    #         "grounding_mode": "Grounded evidence vs synthesis-only",
-    #         "as_of_utc": "As-of timestamp (UTC)",
-    #         "time_scope": "Time horizon for the analysis",
-    #         "target_start": "Range start",
-    #         "target_end": "Range end",
-    #         "domain_id": "Domain preset identifier",
-    #         "domain_guide": "Domain guidance",
-    #         "domain_persona": "Analyst persona",
-    #         "domain_rubric": "Evaluation rubric",
-    #         "domain_format": "Output format",
-    #     },
-    #     param_properties={
-    #         "grounding_mode": {
-    #             "type": "string",
-    #             "enum": ["grounded_required", "synthesis_only"],
-    #         },
-    #         "time_scope": {
-    #             "type": "string",
-    #             "enum": ["current", "historical", "future", "mixed"],
-    #         },
-    #     },
-    # ),
 }
